requirements.txt
Django==3.1.6
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
django-bootstrap4 ==2.3.1
django-cors-headers==3.7.0
opencv-python==4.5.1.48
numpy
celery==5.0
redis==3.5.3